only manage the config for slurmdbd+mysql and all snap.modes
Previously, the snap was generating the slurm.conf in the
configure hook when the snap.mode is set. This was generating
issues when running the snap in distributed mode, where other
tooling needs to manage the slurm.conf.
These changes check the snap mode and only manage the config
if the snap.mode is slurmdbd+mysql or all.

Fixes #93